+++ src/display.js
@@ -141,12 +141,43 @@ function showProjects (content) {
     }
     return taskList.outerHTML
   }
+
+  function projects() {
+    const project = document.getElementById('project');
+    const addProjectButton = document.createElement('button');
+    addProjectButton.classList.add("addButton");
+  
+    project.innerHTML = '<h2>Projects</h2>';
+    addProjectButton.innerHTML = '+';
+    const form = document.createElement('form');
+    const label = document.createElement('label');
+    const input = document.createElement('input');
+  
+    label.setAttribute('for', 'ptitle');
+    input.setAttribute('type', 'text');
+    input.setAttribute('id', 'ptitle');
+    input.setAttribute('name', 'ptitle');
+  
+    addProjectButton.addEventListener('click', () => {
+      add.addProjectToProjects()
+    })
+  
+    form.appendChild(label)
+    form.appendChild(input)
+  
+    project.appendChild(addProjectButton)
+    project.appendChild(form)
+  
+    return project
+  }
+  
  
   
   return{ 
       showProjects,
       showTask,
-      currentTaskList
+      currentTaskList,
+      projects
   }
 })()
