+++ src/pageload.js
@@ -3,40 +3,6 @@
 
 import show from './display'
 
-function projects() {
-  const project = document.getElementById('project');
-  const addProjectButton = document.createElement('button');
-  addProjectButton.classList.add("addButton");
-
-  project.innerHTML = '<h2>Projects</h2>';
-  addProjectButton.innerHTML = '+';
-  const form = document.createElement('form');
-  const label = document.createElement('label');
-  const input = document.createElement('input');
-
-  label.setAttribute('for', 'ptitle');
-  input.setAttribute('type', 'text');
-  input.setAttribute('id', 'ptitle');
-  input.setAttribute('name', 'ptitle');
-
-  addProjectButton.addEventListener('click', () => {
-    addProjectToProjects()
-  })
-
-  form.appendChild(label)
-  form.appendChild(input)
-
-  project.appendChild(addProjectButton)
-  project.appendChild(form)
-
-  return project
-}
-
-
-
-
-
-
 
 
 
@@ -51,7 +17,7 @@ function loadPage () {
     let eId = e.target.id
     show.showTask(eId)
   })
-  projects()
+  show.projects()
   show.showProjects(projectList)
   projectColumn.appendChild(projectList)
